Divide public js files with browserify
