+++ helpers.js
@@ -3,6 +3,5 @@ exports.dump = function(obj) { return JSON.stringify(obj, null, 2) };
 
 exports.handleError = function(err) {
 	if(err) {
-		return res.send('Oops an error occurred', err.message, err['status-code']);
 	}
 }
