+++ package.json
@@ -1,27 +1,28 @@
 {
-  "name": "live-poll",
-  "version": "1.0.0",
-  "description": "",
-  "main": "server.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "nodemon server.js"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "devDependencies": {
-    "browserify": "^16.1.1",
-    "dotenv": "^5.0.1",
-    "nodemon": "^1.17.2",
-    "uglify-js": "^3.3.16",
-    "watchify": "^3.11.0"
-  },
-  "dependencies": {
-    "body-parser": "^1.18.2",
-    "ejs": "^2.5.8",
-    "express": "^4.16.3",
-    "nano": "^6.4.3",
-    "ws": "^5.1.0"
-  }
+	"name": "live-poll",
+	"version": "1.0.0",
+	"description": "",
+	"main": "server.js",
+	"scripts": {
+		"start": "nodemon server.js",
+		"watch": "watchify public/index.js --outfile public/bundle.js & nodemon server.js --inspector",
+		"test": "echo \"Error: no test specified\" && exit 1"
+	},
+	"keywords": [],
+	"author": "",
+	"license": "ISC",
+	"devDependencies": {
+		"browserify": "^16.1.1",
+		"dotenv": "^5.0.1",
+		"nodemon": "^1.17.2",
+		"uglify-js": "^3.3.16",
+		"watchify": "^3.11.0"
+	},
+	"dependencies": {
+		"body-parser": "^1.18.2",
+		"ejs": "^2.5.8",
+		"express": "^4.16.3",
+		"nano": "^6.4.3",
+		"ws": "^5.1.0"
+	}
 }
