+++ public/eventChecker.js
@@ -1,0 +1,8 @@
+exports.addClickEvent = function addClickEvent(el, callback) {
+	if (document.addEventListener) {
+		el.addEventListener('click', callback);
+	} else {
+		el.attachEvent('onclick', callback);
+	}
+}
+
