+++ views/pollDetails.ejs
@@ -25,6 +25,6 @@
 	<input type="submit" value="Vote"/>
 </form>
 
-<script src="/index.js"></script>
+<script src="/bundle.js"></script>
 
 <% include ./partials/footer.ejs %>
